Extract simulation results

This function extracts relevant information from multiple simulations of the same trial specification in a tidy data.frame (1 simulation per row). See also the summary function.

Usage

extract_results(
  object,
  select_strategy = "control if available",
  select_last_arm = FALSE,
  select_preferences = NULL,
  te_comp = NULL,
  raw_ests = FALSE
)

Arguments

object

trial_results object, output from the run_trials function.

select_strategy

single character string. For trials not stopped due to superiority (or with only 1 arm remaining, if select_last_arm is set to TRUE in trial designs with a common control arm; see below), this parameter specifies which arm will be considered selected when calculating trial design performance metrics (described below; this corresponds to the consequence of an inconclusive trial, i.e., which arm would then be used in practice).
The following options are available and must be written exactly as below (case sensitive, cannot be abbreviated):

"control if available" (default): selects the first control arm for trials with a common control arm if this arm is active at end-of-trial, otherwise no arm will be selected. For trial designs without a common control, no arm will be selected.

"none": selects no arm in trials not ending with superiority.

"control": similar to "control if available", but will throw an error for trial designs without a common control arm.

"final control": selects the final control arm regardless of whether the trial was stopped for practical equivalence, futility, or at the maximum sample size; this strategy can only be specified for trial designs with a common control arm.

"control or best": selects the first control arm if still active at end-of-trial, otherwise selects the best remaining arm (defined as the remaining arm with the highest probability of being the best in the final analysis). Only works for trial designs with a common control arm.

"best": selects the best remaining arm (as described under "control or best").

"list or best": selects the first remaining arm from a specified list (specified using select_preferences, technically a character vector). If none of these arms are are active at end-of-trial, the best remaining arm will be selected (as described above).

"list": as specified above, but if no arms on the provided list remain active at end-of-trial, no arm is selected.

select_last_arm

single logical, defaults to FALSE. If TRUE, the only remaining active arm (the last control) will be selected in trials with a common control arm ending with equivalence or futility, before considering the options specified in select_strategy. Must be FALSE for trial designs without a common control arm.

select_preferences

character vector specifying a number of arms used for selection if one of the "list or best" or "list" options are specified for select_strategy. Can only contain valid arms available in the trial.

te_comp

character string, treatment-effect comparator. Can be either NULL (the default) in which case the first control arm is used for trial designs with a common control arm, or a single trial arm. Will be used when calculating sq_err_te (the squared error of the treatment effect comparing the selected arm to the comparator arm, as described below).

raw_ests

single logical. If FALSE (default), the posterior estimates (post_ests, see setup_trial and run_trial) will be used to calculate sq_err (the squared error of the estimated compared to the specified effect in the selected arm) and sq_err_te (the squared error of the treatment effect comparing the selected arm to the comparator arm, as described for te_comp and below). If TRUE, the raw estimates (raw_ests, see setup_trial and run_trial) will be used instead of the posterior estimates.

Value

A data.frame containing the following columns:

sim: the simulation number (from 1 to the number of simulations).

final_n: the final sample size in each simulation.

sum_ys: the sum of the total counts in all arms, e.g., the total number of events in trials with a binary outcome (setup_trial_binom) or the sum of the arm totals in trials with a continuous outcome (setup_trial_norm).

ratio_ys: calculated as sum_ys/final_n.

final_status: the final trial status for each simulation, either "superiority", "equivalence", "futility", or "max", as described in run_trial.

superior_arm: the final superior arm in simulations stopped for superiority, will be NA in simulations not stopped for superiority.

selected_arm: the final selected arm (as described above), will correspond to the superior_arm in simulations stopped for superiority and be NA if no arm is selected. See select_strategy above.

sq_err: the squared error of the estimate in the selected arm, calculated as (estimated effect - true effect)^2 for the selected arms.

sq_err_te: the squared error of the treatment effect comparing the selected arm to the comparator arm (as specified in te_comp). Calculated as:
((estimated effect in the selected arm - estimated effect in the comparator arm) - (true effect in the selected arm - true effect in the comparator arm))^2
Will be NA for simulations without a selected arm or with no comparator specified (see te_comp above).

Examples

# Setup a trial specification
binom_trial <- setup_trial_binom(arms = c("A", "B", "C", "D"),
                                 control = "A",
                                 true_ys = c(0.20, 0.18, 0.22, 0.24),
                                 data_looks = 1:20 * 100)

# Run 10 simulations with a specified random base seed
res <- run_trials(binom_trial, n_rep = 10, base_seed = 12345)

# Extract results and Select the control arm if available
# in simulations not ending with superiority
extract_results(res, select_strategy = "control")
#>    sim final_n sum_ys  ratio_ys final_status superior_arm selected_arm
#> 1    1    2000    428 0.2140000          max         <NA>            A
#> 2    2    2000    384 0.1920000          max         <NA>            A
#> 3    3    1300    252 0.1938462  superiority            B            B
#> 4    4    2000    383 0.1915000          max         <NA>            A
#> 5    5    2000    396 0.1980000          max         <NA>         <NA>
#> 6    6    2000    398 0.1990000          max         <NA>            A
#> 7    7    2000    391 0.1955000          max         <NA>            A
#> 8    8    2000    397 0.1985000          max         <NA>            A
#> 9    9    1700    318 0.1870588  superiority            B            B
#> 10  10     200     39 0.1950000  superiority            B            B
#>          sq_err   sq_err_te
#> 1  8.171730e-05          NA
#> 2  1.744715e-03          NA
#> 3  1.222448e-07 0.020809011
#> 4  5.746768e-04          NA
#> 5            NA          NA
#> 6  4.106669e-04          NA
#> 7  1.666568e-05          NA
#> 8  6.160231e-07          NA
#> 9  2.920080e-04 0.001345626
#> 10 8.576564e-03 0.027148023