Print methods for adaptive trial objects

Prints contents of the first input x in a human-friendly way, see Details for more information.

Usage

# S3 method for trial_spec
print(x, prob_digits = 3, ...)

# S3 method for trial_result
print(x, prob_digits = 3, ...)

# S3 method for trial_results
print(
  x,
  select_strategy = "control if available",
  select_last_arm = FALSE,
  select_preferences = NULL,
  te_comp = NULL,
  raw_ests = FALSE,
  restrict = NULL,
  digits = 1,
  ...
)

# S3 method for trial_results_summary
print(x, digits = 1, ...)

Arguments

x

object to print, see Details below.

prob_digits

single integer, the number of digits used when printing probabilities, allocation probabilities and softening powers.

...

additional arguments, not used.

select_strategy

single character string. For trials not stopped due to superiority (or with only 1 arm remaining, if select_last_arm is set to TRUE in trial designs with a common control arm; see below), this parameter specifies which arm will be considered selected when calculating trial design performance metrics (described below; this corresponds to the consequence of an inconclusive trial, i.e., which arm would then be used in practice).
The following options are available and must be written exactly as below (case sensitive, cannot be abbreviated):

"control if available" (default): selects the first control arm for trials with a common control arm if this arm is active at end-of-trial, otherwise no arm will be selected. For trial designs without a common control, no arm will be selected.

"none": selects no arm in trials not ending with superiority.

"control": similar to "control if available", but will throw an error for trial designs without a common control arm.

"final control": selects the final control arm regardless of whether the trial was stopped for practical equivalence, futility, or at the maximum sample size; this strategy can only be specified for trial designs with a common control arm.

"control or best": selects the first control arm if still active at end-of-trial, otherwise selects the best remaining arm (defined as the remaining arm with the highest probability of being the best in the final analysis). Only works for trial designs with a common control arm.

"best": selects the best remaining arm (as described under "control or best").

"list or best": selects the first remaining arm from a specified list (specified using select_preferences, technically a character vector). If none of these arms are are active at end-of-trial, the best remaining arm will be selected (as described above).

"list": as specified above, but if no arms on the provided list remain active at end-of-trial, no arm is selected.

select_last_arm

single logical, defaults to FALSE. If TRUE, the only remaining active arm (the last control) will be selected in trials with a common control arm ending with equivalence or futility, before considering the options specified in select_strategy. Must be FALSE for trial designs without a common control arm.

select_preferences

character vector specifying a number of arms used for selection if one of the "list or best" or "list" options are specified for select_strategy. Can only contain valid arms available in the trial.

te_comp

character string, treatment-effect comparator. Can be either NULL (the default) in which case the first control arm is used for trial designs with a common control arm, or a single trial arm. Will be used when calculating sq_err_te (the squared error of the treatment effect comparing the selected arm to the comparator arm, as described below).

raw_ests

single logical. If FALSE (default), the posterior estimates (post_ests, see setup_trial() and run_trial()) will be used to calculate sq_err (the squared error of the estimated compared to the specified effect in the selected arm) and sq_err_te (the squared error of the treatment effect comparing the selected arm to the comparator arm, as described for te_comp and below). If TRUE, the raw estimates (raw_ests, see setup_trial() and run_trial()) will be used instead of the posterior estimates.

restrict

single character string or NULL. If NULL (default), results are summarised for all simulations; if "superior, results are summarised for simulations ending with superiority only; if "selected", results are summarised for simulations ending with a selected arm (according to the specified arm selection strategy for simulations not ending with superiority). Some summary measures (e.g., prob_conclusive) can only be calculated across all simulations and several are calculated regardless of restrict settings, but have substantially different interpretations if restricted.

digits

single integer, number of digits to print for probabilities and some other summary values (with 2 extra digits added for outcome rates).

Value

Invisibly returns x.

Details

The behaviour depends on the class of x:

trial_spec: prints a trial specification setup by setup_trial(), setup_trial_binom() or setup_trial_norm().

trial_result: prints the results of a single trial simulated by run_trial(). More details are saved in the trial_result object and thus printed if the sparse argument in run_trial() or run_trials() is set to FALSE; if TRUE, fewer details are printed, but the omitted details are available by printing the trial_spec object created by setup_trial(), setup_trial_binom() or setup_trial_norm().

trial_results: prints the results of multiple simulations generated using run_trials(). Further documentation on how multiple trials are summarised before printing can be found in the summary() function documentation.

trial_results_summary: print method for summary of multiple simulations of the same trial specification, generated by using the summary() function on an object generated by run_trials().

Methods (by class)

trial_spec: Trial specification

trial_result: Single trial result

trial_results: Multiple trial results

trial_results_summary: Summary of multiple trial results