Setup a trial specification using a continuous, normally distributed outcome

Specifies the design of an adaptive trial with a continuous, normally distributed outcome and validates all inputs. Uses normally distributed posterior distributions for the mean values in each trial arm; technically, no priors are used (as using normal-normal conjugate prior models with extremely wide or uniform priors gives similar results for these simple, unadjusted estimates). Technically, this thus corresponds to using improper, flat priors, although not explicitly specified as such. Use run_trial or run_trials to conduct single/multiple simulations of the specified trial, respectively.
Note: add_info as specified in setup_trial is set to the arms and standard deviations used for trials specified using this function.
Further details: please see setup_trial. See setup_trial_binom for simplified setup of trials with binomially distributed binary outcomes.
For additional trial specification examples, see the the Basic examples vignette (vignette("Basic-examples", package = "adaptr")) and the Advanced example vignette (vignette("Advanced-example", package = "adaptr")).

Usage

setup_trial_norm(
  arms,
  true_ys,
  sds,
  start_probs = NULL,
  fixed_probs = NULL,
  min_probs = rep(NA, length(arms)),
  max_probs = rep(NA, length(arms)),
  data_looks = NULL,
  max_n = NULL,
  look_after_every = NULL,
  control = NULL,
  control_prob_fixed = NULL,
  inferiority = 0.01,
  superiority = 0.99,
  equivalence_prob = NULL,
  equivalence_diff = NULL,
  equivalence_only_first = NULL,
  futility_prob = NULL,
  futility_diff = NULL,
  futility_only_first = NULL,
  highest_is_best = FALSE,
  soften_power = 1,
  cri_width = 0.95,
  n_draws = 5000,
  robust = FALSE,
  description = "generic normally distributed outcome trial"
)

Arguments

arms

character vector with unique names for the trial arms.

true_ys

numeric vector, simulated means of the outcome in all trial arms.

sds

numeric vector, true standard deviations (must be > 0) of the outcome in all trial arms.

start_probs

numeric vector, allocation probabilities for each arm at the beginning of the trial. The default (NULL) is automatically changed to equal randomisation.

fixed_probs

numeric vector, fixed allocation probabilities for each arm - must be either a numeric vector with NA for arms without fixed probabilities and values between 0 and 1 for the other arms or NULL (default), if adaptive randomisation is used for all arms or if one of the special settings ("sqrt-based", "sqrt-based start", "sqrt-based fixed", or "match") is specified for control_prob_fixed (described below).

min_probs

numeric vector, lower threshold for adaptive allocation probabilities, lower probabilities will be rounded up to these values. Must be NA (default for all arms) if no boundary is wanted.

max_probs

numeric vector, upper threshold for adaptive allocation probabilities, higher probabilities will be rounded down to these values. Must be NA (default for all arms) if no boundary is wanted.

data_looks

vector of increasing integers, specifies when to conduct adaptive analyses (= the total number of patients randomised at each adaptive analysis). The last number in the vector represents the maximum sample size. Instead of specifying data_looks, the max_n and look_after_every arguments can be used in combination (then data_looks must be NULL, the default).

max_n

single integer, maximum total sample size (defaults to NULL). Must only be specified if data_looks is NULL. Requires specification of the look_after_every argument.

look_after_every

single integer, specified together with max_n. Adaptive analyses will be conducted after every look_after_every patients randomised, and at the total sample size as specified by max_n (max_n does not need to be a multiple of look_after_every). If specified, data_looks must be NULL (as default).

control

single character string, name of one of the arms or NULL (default). If specified, this arm will serve as a common control arm, to which all other arms will be compared and the inferiority/superiority/equivalence thresholds (see below) will be for those comparisons. See setup_trial Details below for information on behaviour with respect to these comparisons.

control_prob_fixed

if a common control arm is specified, this must be set to either NULL (the default), in which case the control arm allocation probability will not be fixed if control arms change (the allocation probability to the first control arm may still be fixed using fixed_probs) Otherwise a vector of probabilities of either length 1 or number of arms - 1 can be provided, or one of the special arguments "sqrt-based", "sqrt-based start", "sqrt-based fixed" or "match". See setup_trial Details below for details in behaviour.

inferiority

single numeric (> 0 and <1, default is 0.01) specifying the inferiority threshold. An arm will be considered inferior and dropped if the probability that it is best (when comparing all arms) or better than the control arm (when a common control is used) drops below this threshold.

superiority

single numeric (>0 and <1, default is 0.99) specifying the superiority threshold. If the probability that an arm is best (when comparing all arms) or better than the control arm (when a common control is used) exceeds this number, said arm will be declared the winner and the trial will be stopped (if no common control is used or if the last comparator is dropped in a design with a common control) or become the new control and the trial will continue (if a common control is specified).

equivalence_prob

single numeric (> 0 and < 1) or NULL (default, corresponding to no equivalence assessment). If a numeric value is specified, arms will be stopped for equivalence if the probability of either (a) equivalence compared to a common control or (b) equivalence between all arms remaining (designs without a common control) exceeds this threshold. Requires specification of equivalence_diff, equivalence_only_first, and a common control arm.

equivalence_diff

single numeric value (> 0) or NULL (default, corresponding to no equivalence assessment). If a numeric value is specified, estimated differences below this threshold will be considered equivalent when assessing equivalence. For designs with a common control arm, the differences between each non-control arm and the control arm is used, and for trials without a common control arm, the difference between the highest and lowest estimated outcome rates are used and the trial is only stopped for equivalence if all remaining arms are thus equivalent.

equivalence_only_first

single logical in trial specifications where equivalence_prob and equivalence_diff are specified, otherwise NULL (default). Must be NULL for designs without a common control arm. If a common control arm is used, this specifies whether equivalence will only be assessed for the first control (if TRUE) or also for subsequent control arms (if FALSE) if one arm is superior to the first control and becomes the new control.

futility_prob

single numeric (> 0 and < 1) or NULL (default, corresponds to no futility assessment). If a numeric value is specified, arms will be stopped for futility when the probability for futility compared to the common control exceeds this threshold. Requires a common control arm, specification of futility_diff and futility_only_first.

futility_diff

single numeric value (> 0) or NULL (default, corresponding to no futility assessment). If a numeric value is specified, estimated differences below this threshold in the beneficial direction (as specified in highest_is_best) will be considered futile when assessing futility in designs with a common control arm. If only 1 arm remains after dropping arms for futility, the trial will be stopped without declaring the last arm superior.

futility_only_first

single logical in trial specifications designs where futility_prob and futility_diff are specified, otherwise NULL (default). Must be NULLfor designs without a commoncontrolarm. Specifies whether futility will only be assessed against the firstcontrol(ifTRUE) or also for subsequent control arms (if FALSE`) if one arm is superior to the first control and becomes the new control.

highest_is_best

single logical, specifies whether larger estimates of the outcome are favourable or not; defaults to FALSE, corresponding to, e.g., an undesirable binary outcomes (e.g., mortality) or a continuous outcome where lower numbers are preferred (e.g., hospital length of stay).

soften_power

either a single numeric value or a numeric vector of exactly the same length as the maximum number of looks/adaptive analyses. Values must be between 0 and 1 (default); if < 1, then re-allocated non-fixed allocation probabilities are all raised to this power to make allocation probabilities less extreme, in turn used to redistribute remaining probability while respecting limits when defined by min_probs and/or max_probs. If 1, then no softening is applied.

cri_width

single numeric >= 0 and < 1, the width of the percentile-based credible intervals used when summarising individual trial results. Defaults to 0.95, corresponding to 95% credible intervals.

n_draws

single integer, the number of draws from the posterior distributions (for each arm) used when running the trial. Defaults to 5000; can be reduced for a speed gain (at the potential loss of stability of results if too low) or increased for increased precision (takes longer). Values < 100 are not allowed and values < 1000 are not recommended and warned against.

robust

single logical, if TRUE (default) the medians and median absolute deviations (scaled to be comparable to the standard deviation for normal distributions; MAD_SD) are used to summarise the posterior distributions; if FALSE, the means and standard deviations (SDs) are used instead (slightly faster, but may be less appropriate for posteriors skewed on the natural scale).

description

character string, default is "generic normally distributed outcome trial". See arguments of setup_trial.

Value

A trial_spec object used to run simulations by run_trial or run_trials. The output is essentially a list containing the input values (some combined in a data.frame called trial_arms), but its class signals that these inputs have been validated and inappropriate combinations and settings have been ruled out. Also contains best_armholding the arm(s) with the best value(s) in true_ys. Use str() to peruse the actual content of the returned object

Details

Because the posteriors used in this type of trial (with a generic, continuous, normally distributed outcome) are by definition normally distributed, FALSE is used as the default value for the robust argument.

Examples

# Setup a trial specification using a continuous, normally distributed, desirable outcome
norm_trial <- setup_trial_norm(
  arms = c("Control", "New A", "New B", "New C"),
  true_ys = c(15, 20, 14, 13),
  sds = c(2, 2.5, 1.9, 1.8), # SDs in each arm
  max_n = 500,
  look_after_every = 50,
  control = "Control", # Common control arm
  # Square-root-based, fixed control group allocation ratios
  control_prob_fixed = "sqrt-based fixed",
  # Desirable outcome
  highest_is_best = TRUE,
  soften_power = 0.5 # Limit extreme allocation ratios
)

# Print using 3 digits for probabilities
print(norm_trial, prob_digits = 3)
#> Trial specification: generic normally distributed outcome trial
#> * Desirable outcome
#> * Common control arm: Control 
#> * Control arm probability fixed at 0.366 (for 4 arms), 0.414 (for 3 arms), 0.5 (for 2 arms)
#> * Best arm: New A
#> 
#> Arms, true outcomes, starting allocation probabilities 
#> and allocation probability limits:
#>     arms true_ys start_probs fixed_probs min_probs max_probs
#>  Control      15       0.366       0.366        NA        NA
#>    New A      20       0.211       0.211        NA        NA
#>    New B      14       0.211       0.211        NA        NA
#>    New C      13       0.211       0.211        NA        NA
#> 
#> Maximum sample size: 500 
#> Maximum number of data looks: 10
#> Planned looks after every 50 patients until maximum sample size
#> Superiority threshold: 0.99 
#> Inferiority threshold: 0.01 
#> No equivalence threshold
#> No futility threshold
#> Soften power for all analyses: 0.5
#> 
#> Additional info: Arm SDs - Control: 2; New A: 2.5; New B: 1.9; New C: 1.8.