Simulate multiple trials

This function conducts multiple simulations using a trial specification as specified by setup_trial, setup_trial_binom or setup_trial_norm. This function essentially manages random seeds and runs multiple simulation using run_trial - additional details on individual simulations are provided in that function's description. This function allows simulating trials in parallel using multiple cores, automatically saving and re-loading saved objects, and "growing" already saved simulation files (i.e., appending additional simulations to the same file).

Usage

run_trials(
  trial_spec,
  n_rep,
  path = NULL,
  overwrite = FALSE,
  grow = FALSE,
  cores = 1,
  base_seed = NULL,
  sparse = TRUE,
  progress = NULL,
  version = NULL,
  compress = TRUE
)

Arguments

trial_spec

trial_spec object, generated and validated by the setup_trial, setup_trial_binom or setup_trial_norm function.

n_rep

single integer; the number of simulations to run.

path

single character; if specified (defaults to NULL), files will be written to and loaded from this path using the saveRDS / readRDS functions.

overwrite

single logical; defaults to FALSE, in which case previous simulations saved in the same path will be re-loaded (if the same trial specification was used). If TRUE, the previous file is overwritten. If grow is TRUE, this argument must be set to FALSE.

grow

single logical; defaults to FALSE. If TRUE and a valid path to a valid previous file containing less simulations than n_rep, the additional number of simulations will be run (appropriately re-using the same base_seed, if specified) and appended to the same file.

cores

single integer; the number of cores to run the simulations on using the parallel library. Defaults to 1; may be increased to run multiple simulations in parallel. parallel::detectCores() may be used to find the number of available cores.

base_seed

single integer; a random seed used as the basis for simulations; each simulation will set the random seed to a value based on this (+ the trial number), without affecting the global random seed after the function has been run.

sparse

single logical, as described in run_trial; defaults to TRUE when running multiple simulations, in which case only the data necessary to summarise all simulations are saved for each simulation. If FALSE, more detailed data for each simulation is saved, allowing more detailed printing of individual trial results and plotting using plot_history (plot_status does not require non-sparse results).

progress

single numeric > 0 and <= 1 or NULL. If NULL (default), no progress is printed to the console. Otherwise, progress messages are printed to the control at intervals proportional to the value specified by progress.
Note: as printing is not possible from within clusters on multiple cores, the function conducts batches of simulations on multiple cores (if specified), with intermittent printing of statuses. Thus, all cores have to finish running their current assigned batches before the other cores may proceed with the next batch. If there is substantial differences in the simulation speeds across cores, using progress may thus increase total simulation times.

version

passed to saveRDS when saving simulations, defaults to NULL (as in saveRDS), which means that the current default version is used. Ignored if simulations are not saved.

compress

passed to saveRDS when saving simulations, defaults to TRUE (as in saveRDS), see saveRDS for other options. Ignored if simulations are not saved.

Value

A list of a special class "trial_results", which contains the trial_results (results from all simulations), trial_spec (the trial specification), n_rep, base_seed, elapsed_time (the total simulation run time) and sparse (as described above). These results may be extracted using the extract_results function and summarised using the summary or print (print.trial_results) functions; see function documentation for details on additional arguments used to select arms in simulations not ending in superiority and other summary choices.

Examples

# Setup a trial specification
binom_trial <- setup_trial_binom(arms = c("A", "B", "C", "D"),
                                 true_ys = c(0.20, 0.18, 0.22, 0.24),
                                 data_looks = 1:20 * 100)

# Run 10 simulations with a specified random base seed
res <- run_trials(binom_trial, n_rep = 10, base_seed = 12345)

# See ?summary and ?print for details on summarising and printing