Summary of simulated trial results

Summarises simulation results from the run_trials function. Uses extract_results, which may be used directly to extract key trial results without summarising.

Usage

# S3 method for trial_results
summary(
  object,
  select_strategy = "control if available",
  select_last_arm = FALSE,
  select_preferences = NULL,
  te_comp = NULL,
  raw_ests = FALSE,
  restrict = NULL,
  ...
)

Arguments

object

trial_results object, output from the run_trials function.

select_strategy

single character string. For trials not stopped due to superiority (or with only 1 arm remaining, if select_last_arm is set to TRUE in trial designs with a common control arm; see below), this parameter specifies which arm will be considered selected when calculating trial design performance metrics (described below; this corresponds to the consequence of an inconclusive trial, i.e., which arm would then be used in practice).
The following options are available and must be written exactly as below (case sensitive, cannot be abbreviated):

"control if available" (default): selects the first control arm for trials with a common control arm if this arm is active at end-of-trial, otherwise no arm will be selected. For trial designs without a common control, no arm will be selected.

"none": selects no arm in trials not ending with superiority.

"control": similar to "control if available", but will throw an error for trial designs without a common control arm.

"final control": selects the final control arm regardless of whether the trial was stopped for practical equivalence, futility, or at the maximum sample size; this strategy can only be specified for trial designs with a common control arm.

"control or best": selects the first control arm if still active at end-of-trial, otherwise selects the best remaining arm (defined as the remaining arm with the highest probability of being the best in the final analysis). Only works for trial designs with a common control arm.

"best": selects the best remaining arm (as described under "control or best").

"list or best": selects the first remaining arm from a specified list (specified using select_preferences, technically a character vector). If none of these arms are are active at end-of-trial, the best remaining arm will be selected (as described above).

"list": as specified above, but if no arms on the provided list remain active at end-of-trial, no arm is selected.

select_last_arm

single logical, defaults to FALSE. If TRUE, the only remaining active arm (the last control) will be selected in trials with a common control arm ending with equivalence or futility, before considering the options specified in select_strategy. Must be FALSE for trial designs without a common control arm.

select_preferences

character vector specifying a number of arms used for selection if one of the "list or best" or "list" options are specified for select_strategy. Can only contain valid arms available in the trial.

te_comp

character string, treatment-effect comparator. Can be either NULL (the default) in which case the first control arm is used for trial designs with a common control arm, or a single trial arm. Will be used when calculating sq_err_te (the squared error of the treatment effect comparing the selected arm to the comparator arm, as described below).

raw_ests

single logical. If FALSE (default), the posterior estimates (post_ests, see setup_trial and run_trial) will be used to calculate sq_err (the squared error of the estimated compared to the specified effect in the selected arm) and sq_err_te (the squared error of the treatment effect comparing the selected arm to the comparator arm, as described for te_comp and below). If TRUE, the raw estimates (raw_ests, see setup_trial and run_trial) will be used instead of the posterior estimates.

restrict

single character string or NULL. If NULL (default), results are summarised for all simulations; if "superior, results are summarised for simulations ending with superiority only; if "selected", results are summarised for simulations ending with a selected arm (according to the specified arm selection strategy for simulations not ending with superiority). Some summary measures (e.g., prob_conclusive) can only be calculated across all simulations and several are calculated regardless of restrict settings, but have substantially different interpretations if restricted.

...

additional arguments, not used.

Value

A "trial_results_summary" object containing the following:

n_rep: the number of simulations.

n_summarised: the number of simulations summarised.

highest_is_best: as specified in setup_trial.

elapsed_time: the total simulation time.

size_mean, size_sd, size_median, size_p25, size_p75: the mean, standard deviation, median as well as 25- and 75-percentiles of the sample sizes of the summarised trial simulations.

sum_ys_mean, sum_ys_sd, sum_ys_median, sum_ys_p25, sum_ys_p75: the mean, standard deviation, median as well as 25- and 75-percentiles of the total sum_ys (e.g., the total number of events in trials with a binary outcome, or the sums of continuous values for all patients across all arms in trials with a continuous outcome) across all arms in the summarised trial simulations.

ratio_ys_mean, ratio_ys_sd, ratio_ys_median, ratio_ys_p25, ratio_ys_p75: the mean, standard deviation, median as well as 25- and 75-percentiles of the final ratio_ys (sum_ys/final_n) across all arms in the summarised trial simulations.

prob_conclusive: the proportion of conclusive trial simulations (simulations not stopped at the maximum sample size without a superiority, equivalence or futility decision).

prob_superior, prob_equivalence, prob_futility, prob_max: the proportion (0-1) of trial simulations stopped for superiority, equivalence, futility or inconclusive at the maximum allowed sample size, respectively.

prob_select_*: the selection probabilities for each arm and for no selection, according to the specified selection strategy. Contains one element per arm, named as prob_select_arm_<arm name> and prob_select_none for the probability of selecting no arm.

rmse, rmse_te: the root mean squared error of the estimates for the selected arm and for the treatment effect, as described further in extract_results.

idp: the ideal design percentage (IDP; 0-100%), see Details.

select_strategy, select_last_arm, select_preferences, te_comp, raw_ests,restrict: as specified above.

control: the control arm specified by setup_trial, setup_trial_binom or setup_trial_norm; NULL if no control.

equivalence_assessed, futility_assessed: single logicals, specifies whether the trial design specification includes assessments of equivalence and/or futility.

base_seed: as specified in run_trials.

cri_width, n_draws, robust, description, add_info: as specified in setup_trial, setup_trial_binom or setup_trial_norm.

Details

The ideal design percentage (IDP) returned (described below) is based on Viele et al, 2020 doi: 10.1177/1740774519877836 and has been adapted to work for trials with both desirable/undesirable outcomes and non-binary outcomes. Briefly, the expected outcome is calculated as the sum of the true outcomes in each arm multiplied by the corresponding selection probabilities (ignoring simulations with no selected arm). The IDP is then calculated as:

For desirable outcomes:
100 * (expected outcome - lowest true outcome) / (highest true outcome - lowest true outcome)

For undesirable outcomes:
100 - IDP calculated for desirable outcomes

Examples

# Setup a trial specification
binom_trial <- setup_trial_binom(arms = c("A", "B", "C", "D"),
                                 control = "A",
                                 true_ys = c(0.20, 0.18, 0.22, 0.24),
                                 data_looks = 1:20 * 100)

# Run 10 simulations with a specified random base seed
res <- run_trials(binom_trial, n_rep = 10, base_seed = 12345)

# Summarise simulations - select the control arm if available in trials not
# ending with a superiority decision
res_sum <- summary(res, select_strategy = "control")

# Print summary
print(res_sum, digits = 1)
#> Multiple simulation results: generic binomially distributed outcome trial
#> * Undesirable outcome
#> * Number of simulations: 10
#> * Number of simulations summarised: 10 (all trials)
#> * Common control arm: A
#> * Selection strategy: first control if available (otherwise no selection)
#> * Treatment effect compared to: no comparison
#> 
#> Performance metrics (using posterior estimates):
#> * Sample sizes: mean 1720.0 (SD: 580.8) | median 2000.0 (IQR: 1775.0 to 2000.0)
#> * Total summarised outcomes: mean 338.6 (SD: 116.8) | median 387.5 (IQR: 334.2 to 396.8)
#> * Total summarised outcome rates: mean 0.196 (SD: 0.007) | median 0.195 (IQR: 0.192 to 0.198)
#> * Conclusive: 30.0%
#> * Superiority: 30.0%
#> * Equivalence: 0.0% [not assessed]
#> * Futility: 0.0% [not assessed]
#> * Inconclusive at max sample size: 70.0%
#> * Selection probabilities: A: 60.0% | B: 30.0% | C: 0.0% | D: 0.0% | None: 10.0%
#> * RMSE: 0.03605
#> * RMSE treatment effect: 0.12820
#> * Ideal design percentage: 77.8%
#> 
#> Simulation details:
#> * Simulation time: 0.603 secs
#> * Base random seed: 12345
#> * Credible interval width: 95%
#> * Number of posterior draws: 5000
#> * Estimation method: posterior medians with MAD-SDs